elfloader: fix variable prototype and remove var
shadowing.

The variable "dtb_size" is of type "size_t" and is defined in
src/arch-arm/sys_boot.c, line 36.

"size_t" is most certainly NOT the same size as "uint32_t", even on
32-bit architectures. Thus, the declaration in smp_boot.c is incorrect,
since it does not match the definition in sys_boot.c. Why even create
a local declaration - put this in a common header file and you will
see those problems right away. Single point of maintenance!

This may lead to an incorrectly sized memory access, that only happens
to be correct by chance in Little-Endian mode. For ARM in Big-Endian
mode this is a bug and will most likely result in an incorrect DTB
size of zero.

This fixes c5735119 ("elfloader: pass DTB from bootloader to seL4 on
ARM").

Moreover, remove the shadowing of global variables by defining local
ones with the same name => Rename the local one in src/common.c.

This could have been detected with "-Wshadow".

Practically speaking our DTBs are always (a lot) smaller than 32-bit.
Thus, continue to pass a 32-bit size to the kernel in order to not
change the API here.

+++ elfloader-tool/src/arch-arm/smp_boot.c
@@ -24,7 +24,7 @@ static volatile int non_boot_lock = 0;
 void arm_disable_dcaches(void);
 
 extern void const *dtb;
-extern uint32_t dtb_size;
+extern size_t dtb_size;
 
 /* Entry point for all CPUs other than the initial. */
 void non_boot_main(void)
@@ -69,10 +69,10 @@ void non_boot_main(void)
         arm_enable_mmu();
     }
 
-    /* Jump to the kernel. */
+    /* Jump to the kernel. Note: Our DTB is smaller than 4 GiB. */
     ((init_arm_kernel_t)kernel_info.virt_entry)(user_info.phys_region_start,
                                                 user_info.phys_region_end, user_info.phys_virt_offset,
-                                                user_info.virt_entry, (paddr_t)dtb, dtb_size);
+                                                user_info.virt_entry, (paddr_t)dtb, (uint32_t)dtb_size);
 
     printf("AP Kernel returned back to the elf-loader.\n");
     abort();
