+++ elfloader-tool/src/arch-arm/sys_boot.c
@@ -242,12 +242,13 @@ void continue_boot(int was_relocated)
         printf("Jumping to kernel-image entry point...\n\n");
     }
 
+    /* Jump to the kernel. Note: Our DTB is smaller than 4 GiB. */
     ((init_arm_kernel_t)kernel_info.virt_entry)(user_info.phys_region_start,
                                                 user_info.phys_region_end,
                                                 user_info.phys_virt_offset,
                                                 user_info.virt_entry,
                                                 (word_t)dtb,
-                                                dtb_size);
+                                                (uint32_t)dtb_size);
 
     /* We should never get here. */
     printf("ERROR: Kernel returned back to the ELF Loader\n");
